Fix code to make it compileable with Android NDR r21

+++ libretro/core/libretro-core.cpp
@@ -268,7 +268,10 @@ void retro_get_system_info(struct retro_system_info *info)
 
 void retro_get_system_av_info(struct retro_system_av_info *info)
 {
-   struct retro_game_geometry geom = { retrow, retroh, 1024, 1024,4.0 / 3.0 };
+  struct retro_game_geometry geom = {
+				     (unsigned int) retrow,
+				     (unsigned int) retroh,
+				     1024, 1024,4.0 / 3.0 };
    struct retro_system_timing timing = { 50.0, 44100.0 };
 
    info->geometry = geom;
